Corrección de crear tickets y modificar stock

+++ Floristeria/src/main/java/FlowerStore/Tickets/InvoiceLine.java
@@ -5,9 +5,9 @@ public class InvoiceLine {
     private int idInvoiceHeader;
     private int idProductInvoiceLine;
     private int productQuantity;
-    private double priceInvoiceLine;
+    private float priceInvoiceLine;
 
-    public InvoiceLine( int idInvoiceHeader, int idProductInvoiceLine, int productQuantity, double priceInvoiceLine) {
+    public InvoiceLine( int idInvoiceHeader, int idProductInvoiceLine, int productQuantity, float priceInvoiceLine) {
         this.idInvoiceHeader = idInvoiceHeader;
         this.idProductInvoiceLine = idProductInvoiceLine;
         this.priceInvoiceLine = priceInvoiceLine;
@@ -46,11 +46,11 @@ public class InvoiceLine {
         this.productQuantity = productQuantity;
     }
 
-    public double getPriceInvoiceLine() {
+    public float getPriceInvoiceLine() {
         return priceInvoiceLine;
     }
 
-    public void setPriceInvoiceLine(double priceInvoiceLine) {
+    public void setPriceInvoiceLine(float priceInvoiceLine) {
         this.priceInvoiceLine = priceInvoiceLine;
     }
